cleanup of tests.
